requirements.txt
#autogen
#autogen-agentchat~=0.2
#python-dotenv


autogen-agentchat~=0.2
python-dotenv
pydantic
docker
